requirements.txt
bcrypt==4.1.2
Flask==3.0.2
fuzzywuzzy==0.18.0
jsonify
Flask-SQLAlchemy==3.0.4
Flask-Bcrypt==1.0.1
python-dotenv
